requirements.txt
PyQt5==5.15.11
groq==0.4.1
SpeechRecognition==3.10.4
pyttsx3==2.99
setuptools
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
SQLAlchemy==2.0.36
reportlab==4.0.7
unidecode==1.3.6
